refactor... hopefully for readability
