+++ src/tokenizer.rs
@@ -1,6 +1,5 @@
 mod dict_reader;
 pub mod newmm;
-mod tcc;
-pub(crate) mod tcc_rules;
+pub(crate) mod tcc;
 pub mod tokenizer_trait;
 mod trie_char;
